requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.31.0
apscheduler==3.10.4
streamlit==1.32.2
plotly==5.20.0
pandas==2.2.1
